backend/requirements.txt
fastapi
uvicorn
langchain
openai
pydantic
pydantic-settings
pdfplumber
python-docx
